Remove a2a metadata from sdk
